# website/src/data/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "wallhack REST API",
		"version": "0.3.0",
		"description": "Headless control plane for wallhack nodes. All endpoints except /health require Basic authentication when --api-user and --api-secret are configured."
	},
	"servers": [
		{
			"url": "https://127.0.0.1:6564",
			"description": "Local node (default port)"
		}
	],
	"components": {
		"securitySchemes": {
			"basicAuth": {
				"type": "http",
				"scheme": "basic",
				"description": "Configured via --api-user and --api-secret. Optional — if not configured, no authentication is required."
			}
		},
		"schemas": {
...
			"RouteAddRequest": {
				"type": "object",
				"required": ["cidr", "peer"],
				"properties": {
					"cidr": {
						"type": "string",
						"description": "Target network in CIDR notation (e.g., 10.0.0.0/24)."
					},
					"peer": {
						"type": "string",
						"description": "Name of the peer through which the network should be routed."
					}
				}
			},
			"RouteResponse": {
				"type": "object",
				"required": ["cidr", "peer", "create_time"],
				"properties": {
					"cidr": {
						"type": "string",
						"description": "Network range routed through the tunnel."
					},
					"peer": {
						"type": "string",
						"description": "Name of the peer responsible for this route."
					},
					"create_time": {
						"type": "string",
						"format": "date-time",
						"description": "ISO 8601 UTC timestamp when the route was established."
					}
				}
			},
			"RoutesResponse": {
				"type": "object",
				"required": ["routes"],
				"properties": {
					"routes": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/RouteResponse" },
						"description": "List of active routing entries."
					}
				}
			},
			"SuccessResponse": {
				"type": "object",
				"required": ["success"],
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the requested operation completed successfully."
					},
					"message": {
						"type": "string",
						"description": "Optional detail providing context for success or failure."
					}
				}
			}
		}
	},
	"paths": {
...
		"/peers/{id}": {
			"delete": {
				"summary": "Disconnect peer",
				"description": "Terminates the connection with a specific peer. Accepts the peer's unique id (from GET /peers) or an unambiguous name prefix.",
				"operationId": "disconnectPeer",
				"security": [{ "basicAuth": [] }],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": { "type": "string" },
						"description": "Peer id (or unambiguous name prefix) to disconnect."
					}
				],
				"responses": {
					"200": {
						"description": "Peer disconnected.",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/SuccessResponse" }
							}
						}
					},
					"400": { "description": "Malformed peer name." },
					"401": { "description": "Unauthorized." },
					"404": { "description": "Specified peer not found." }
				}
			}
		},
		"/routes": {
			"get": {
				"summary": "List routing table",
				"description": "Retrieves all active routing entries established through this node.",
				"operationId": "listRoutes",
				"security": [{ "basicAuth": [] }],
				"responses": {
					"200": {
						"description": "Routing table retrieved.",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/RoutesResponse" }
							}
						}
					},
					"401": { "description": "Unauthorized." }
				}
			},
			"post": {
				"summary": "Add route",
				"description": "Configures a new route to a target network through an active peer.",
				"operationId": "addRoute",
				"security": [{ "basicAuth": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/RouteAddRequest" }
						}
					}
				},
				"responses": {
					"201": {
						"description": "Route established.",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/SuccessResponse" }
							}
						}
					},
					"400": {
						"description": "Invalid network specification or peer name."
					},
					"401": { "description": "Unauthorized." }
				}
			}
		},
...
	}
}